pyproject.toml
[tool.poetry]
name = "docsy"
version = "0.1.0"
description = "Docsy creates pull requests on documentation repositories."
authors = ["Felix Zieger <mail@felixzieger.de>"]
readme = "README.md"

[tool.poetry.scripts]
slack = "docsy.slack.slack:flask_app"
github = "docsy.github.github:flask_app"
dashboard = "docsy.dashboard.dashboard:flask_app"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
flask = "^3.0.3"
gunicorn = "^23.0.0"
loguru = "^0.7.2"
playwright = "^1.48.0"
xdg-base-dirs = "^6.0.2"
flask-cors = "^5.0.0"
pandoc = "^2.4"

[tool.poetry.group.engine.dependencies]
openai = "^1.47.0"
aider-chat = "^0.60.1"
gitpython = "^3.1.43"
pygithub = "^2.3.0"
alembic = "^1.13.2"
sqlalchemy = "^2.0.31"

[tool.poetry.group.dashboard.dependencies]
dash = "^2.17.1"
pandas = "^2.2.2"

[tool.poetry.group.slack.dependencies]
slack-bolt = "^1.18.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
